Added a capability to MainWin to syncCutBuffer with the application.  The VNC Java client uses the cutbuffer instead of the selection for Copy/Paste to desktop, and this mechanism keeps the two in sync so Copy/Paste works properly.  Also, added Cut/Copy/Paste menus to the right mouse button of various widgets.
Fixed 3D plotting to work with the vtkCutter so it works
better.  Also, added support for 3D meshes in addition
to clouds.  Meshes store connectivity, so they are better
at representing holes in data.  Fixed the 3D plotter so
that rotate is more intuitive, and added lights so you can
see your data better at any angle.

Fixed the loader so that it can load elements with the ""
value, and so that it doesn't duplicate entries found
more than once by *.xml pattern matching.