Fixed the rendering of groups, and groups within groups. If groups are mixed in with other elements, then they are drawn with a gray outline/heading, with the title taken from the <group><about><label>.  However, if a group contains only other groups, then it is treated as a tabbed notebook, and each group within is put on a separate page.
WARNING: There are many bad interactions between the
blt::tabset, the Rappture::Scroller, and the Rappture::Pager.
Pages shake violently when all are in play.  The only way I
could get them to settle down was by putting the tabs above
the pages they control.  Have to revisit this some time to
make it look better...