fixed the python rpunits setup script so now it really finds PyRpUnits.cc
added more conversions to RpUnitsStd.[h,cc]

added static convert function to RpUnits.[h,cc] so now users can ask for a 
conversion as without having to make an RpUnits Object. instead, the
user specifies the value to be converted (value with attached units) as well
as the units we should convert to and if they want units in the returned
string. if the units exist, the conversion is done, if not, the original 
string should be returned.

also added the ability to load preset units into the dictionary.
use the static function RpUnits::addPresets(...) in c++ this is done
automatically in python (although, maybe it shouldn't be).