Fixed "copy/paste with desktop" to convert any \r's to \n's. This was a problem when pasting from Windows or Mac environments, when what looked like a clean text file would paste badly into a text area.
Fixed the Rappture::exec to cleanly separate stdout/stderr
channels.  We were getting an error in Spice when a stderr
statement appeared right in the middle of a stdout line of data.
The line was truncated, so the value at the end was interpreted
as a very different number, causing a spike on the graph.
(Example: Common Collector Amplifier, and ".print ac im(vs)")

Fixed Field objects to respect the length scale set by the
overall structure.