Fixed the way that tools are launched so that any "MiddlewareTime:" statements that appear in the stdout from the underlying tool process are extracted and fed up to the stderr of the overall Rappture job. This allows us to take information from jobs farmed out to the grid and report the information back up through Rappture to the middleware.
Start times reported by the underlying jobs are automatically offset
by the start time of the tool, and job numbers are offset by the job
number of the tool.  So the tool should report job numbers and start
times relative to its own start, and Rappture will handle the rest.