Fix for support ticket #1455, "floating-point value too large to represent". The user was entering a really big value like "1e400" into the doping area of PN Junction Lab.  The GUI would choke and the job would later fail.  Fixed the field editor to test for bad values and honor min/max limits set on parameters in a device structure.
Fix for support ticket #1465, "bad option "add": should be one of...".
The new molecule viewer requires add/delete operations.  This is now
part of the general structure viewer, but the DeviceViewer1D class was
not updated properly.  It is now.  Fixed the structure demo to copy
the input to the output so we can test for this.