Fixed the Rappture::encoding::encode and decode operations to honor "--". This is handy when the string may start with a dash, as in:     Rappture::encoding::encode -- -hi
Fixed the library.tcl code to guard against encode/decode problem by
using the "--" in front of strings obtained from the xml.

Cleaned up the tests for the Rappture::LibraryObj class.  Everything
works properly now, except for units.  Some of the tests are failing
because the units code reports "length" instead of "area" (test finds
this problem).  This should be fixed in the units code.