enabling user to switch the text in the "Simulate" Button for cceHUB by using the tool.action.label tag
adding examples/lang directory where we can put example programs for each language showing off usage of bindings. most of the code is already in the wiki, but this will allow the user to easily run the example and see the exact text/steps needed to use a binding function.