adding flowvisviewer widget and flowvis-test program for exercising flow commands. most stuff works in the flowvis-test gui, there are not transfer functions or legends. play/pause, stop and record work, record download's or saves the video file to users's desktop. adding iconc for pause, record and stop.
added "flow play" command to nanovis. the goal of this command is to send back
the next frame to the client.  the play functionality allows users to see
frames as if they were a movie, and also send command nanovis server. this
makes the experience more interactive than waiting for a movie to be made and
downloading it. also the play functionality can be used as a cheap preview to
get the best angle and settings before recording the movie.

added sendDataToClient function to nanovis server so i could send back movie
files.  we create the movie on the render server and send back the mpeg binary
for the user to download. the function is pretty general so it can be used to
send any blob of data.