Begin reorganizing Rappture C++ API, starting with the use of Rappture::Outcome.
Instead of returning an Outcome and expecting every routine to
save it, pass it as a reference.  Let's strip it down to what's
needed right now and add back functionality as required.  

Right now too many return values are ignored in the rappture library.
We need to check all results and return the error messages via
the Outcome.  At the topmost level where is touches the developer 
API we can drop the error checking or turn it on via an environment
variable. 

Example. Not enough checks are made for memory allocation failure.  
If the application is running on an over-committed server memory
allocation errors will be silently passed on.  It's okay to be
fault tolerant where possible, but if we fail to check in the internal
library functions, it's too late.

--gah