sequenceresult.tcl:
Playing a sequence continuously and then clicking "Clear" will generate an
infinite number of automatic error 'invalid command name ""' messages.  

   _fixValue    Check for _topmost (top most data object) to be "". 
	        Deletes the viewer.

   delete       Calls pause method to stop the player.



balloon.tcl:    

Bandaid to fix 'bad geometry specifier "-44x184+864+247"' errors.  This has to
do with the placement of the popup menu when clicking on axes in the xyresult
widget.  Added a test if the width or height is negative and clip the geometry
to +x+y.

xyresult.tcl:    

Graph issues 'can't find element "elem1"' error when attempting to 
to deactivate an element that doesn't exist.  The highlighted element name 
is stored in the variable _hilite(elem). It goes out of sync with the graph.
Test for the existence of the element in the graph explicitly.