Fixed support ticket #5831, which produced an error message like:   can't set _cntlInfo(...): bad value "1e+18/cm3"
The problem arose when you simulated two values like 1e+18 and 1.0e+18.
The Radiodial understood that these were the same value, but it honored
only one of the string representations and rejected the other when you
tried to adjust the dials.  It now keeps all recognized string values,
and maps the string onto the real value when you're manipulating the
control.  Also fixed the ResultSet to avoid creating the Radiodial when
the value is numerically the same even though strings are different.