Fixed the image object to act more like a true input.  If an image has a label, it will show a thumbnail along with image info, and users can right-click to get upload/download options.  If there's no label, then the image is shown full-size as before, so it acts like a decoration.
Changed binary string values to look like the image, with an icon representing
binary data and an info string showing info about the data type.  This is
a little nicer than the hex dump we used to show for binary strings.  String
size can be set to "binary" to force the string into the binary display
mode.