Initial version of the "instant rappture" gui builder.  To run this, just say "irappture" or "irappture -tool path/to/tool.xml".  It will bring up an editor that lets you specify inputs/outputs and preview the tool.
Made two fixes along the way to the scroller, so that it handles
automatic scrollbars better, and to the grab stack, so that it avoids
pushing the same window on the grab stack twice.  Both of these fixes
solved problems in iRappture, and they'll probably fix strange behaviors
in the main rappture as well.