Added preview of outputs to the irappture builder.  Fixed a problem with the file chooser for the "save as" dialog.  Added a distinction between errors and warnings when looking for problems in the tool definition.
Fixed a few problems in RapptureGUI affecting the scroller and the
way it refreshes itself when the frame within it suddenly shrinks.
Fixed the histogram to avoid errors when the histogram data is
empty.  Fixed the analyzer so you can create it without setting
the -notebookpage option.  All of these errors showed up in the
irappture builder.