More changes for the regression tester tool.  Added a way for all ObjVal objects to import their attribute values.  Once an ObjVal has been imported, it knows everything about itself without having to consult the XML.
Added Rappture::objects::viewer, which can be used to query the viewer for
a particular object type.  Each object can have an "input" viewer for
specifying input values, and an "output" viewer for visualizing output
results.  Right now, there is only one viewer implemented: an output viewer
for the CurveValue object.  We should create output viewers for all object
types, and then work on the input side too.

Added a "mkobjects.tcl" script used by the Makefile to build up the tclIndex
file.  It looks for input/output viewers for all object classes and adds
them to tclIndex so they can be autoloaded.