Major rewrite of the Analyzer and management of results so that "Clear One Result" will work properly--even with "analyzer last" and "control manual-resim".  The old ResultSet widget is now a ResultSelector widget that takes a ResultSet as a data object.  The ResultSet contains all Rappture::LibraryObj result objects and computes diffs between results. The ResultSet owns all objects; when a result is cleared from the ResultSet, the object is deleted.  The ResultSet notifies various clients (ResultSelector and Analyzer) whenever there are changes, such as adding or clearing a result.
Fixed the core (scew) library to avoid ignoring whitespace and trimming
strings.  This was causing problems with string values and MIME-encoded
images with trailing newlines.  The original (with newlines) seemed like
a different value than the version in the run.xml file (without newlines).
Some simulators may depend on the newlines.  The Tcl library version
preserves the newlines, so the scew version should too.