Fixed the main program to catch kill signals and program termination and flush the log buffer before exiting.  Also saves "finished" as the last message in the log.
Fixed the logger to see if the log directory is missing and warn about
it and turn off logging (rather than crapping out or trying to create
the log directory).  The specific date for the log directory (e.g.,
/data/userlogs/2012-10-19) must be created by the middleware so that all
users can share it.

Added log messages for tooltips on the output quantities, and for almost
everything in the HeightmapViewer.