Fixes for nanoHUB ticket #258058 and #258045 -- note within a group tab doesn't fill out the area properly.  Should work better now.  Also fixed tabs to contain the window, instead of floating above the cotents as they have for a long time.
Fixed boolean controls to have a more obvious on/off switch instead of a
checkbox.  Fixed integers and spinners to use larger +/- buttons that are
easier to press on an iPad.  Fixed numbers and other gauges to have the
same relief style as entries and other widgets.

Added new layout styles to groups:  horizontal, vertical, tabs, and sentence.
You can now explicitly make a layout vertical instead of tabs by setting
the layout to "vertical" instead of throwing in a separator.  Updated the
zoo/groups example to show off new horizontal and sentence types.

Fixed the "drawing" example in the zoo to gray out the trapezoid top when
it is disabled.