three major features being added: 1. setting custom min/max range from the IsosurfaceTab 2. setting custom min/max range from the legend 3. isosurface probe from the legend
Most of the new work was performed in the following methods:
LegendB1Motion
LegendPointToValue
LegendRangeAction
LegendTitleAction
LegendProbeSingleContour

The min/max custom range values are stored in _settings array. Through out
the code, when we want the limits of a field being displayed, we also check
to see if a custom range has been enabled. If it has, in most cases, we use
the custom range as the limits instead of the limits of the currently
displayed field.

The current implementation allows for the custom range to exceed the range of
the data in the field.

also renamed the Combo method to LegendTitleAction to make its objective more clear

split some commonly used features into their own methods, hopefully we can find
common code among all of the viewers and move these to a super class:

DisableMouseRotationBindings
SetupMouseRotationBindings
SetupMousePanningBindings
SetupKeyboardBindins