adding a -validatecommand to gauge.tcl so people can add their own validation to the widget
using gauge widgets for the min/max fields in the settings tab of the
isosurface viewer instead of entry widgets so we can take advantage of input
validation and tooltip cues. tied the min/max editor values, that are above and
below the legend colormap, to the min/max gauge widgets so they all change at
the same time and read the same values. I think there is still a problem if the
min equals the max value, the legend will still show the same value above and
below the colormap. there may also be a race condition when first setting the
min and max gauge widget values because validation occurs. if the data's min is
greater than the default max of 1 (not sure where this comes from yet) or if
the data max is less than the default min of 0 (not sure where this comes from
either) then an error will be raised.