in the vtk isosurface viewer, fixing bugs related to trying to set the min and max gauges at the same time with input validation.
included a SetMinMaxGauges method to handle figuring out if we should set the
min value first or the max value first to avoid input validation errors. also
added the ToggleCustomRange method to help handle syncing the min/max text in
the legend around the colormap and the gauge widgets in the Isosurface Settings
tab. the gauges can lag in being updated with switching between viewing fields.
added -customrangemin and -customrangemax temp variables to the settings array
to hold a backup of the custom range values set by the user, so they don't get
deleted after we turn off the -customrange flag.

removed checks to see if the custom range was within the range of the data when
the following methods are called:

AdjustSetting -field
AdjustSetting -range

I couldn't figure out why the checks were needed.