removing the _ignoreRangeUpdate because newer implementations of the min/max range syncing does not cause a loop between changing the legend values and the gauge values.
added a -range flag to InitSettings in ::Rebuild so the custom ranges are set.i dont think this works fully because some of the adjustments are inside the ToggleCustomRange method, which isn't called by AdjustSetting -range.

remove the trace on the crange widget's textvariable in favor of just calling the ToggleCustomRange method where it needs to be called.

tried using the crange widget's (checkbutton) invoke command to turn on the checkbutton and call its -command script. the invoke command is a toggle, not just a "turn on" command and doesn't work the same as select.

changed the order of turning on the custom range checkbutton and setting the min or max gauge value. we should first signal that the custom range is being enabled from the legend min/max editor, then adjust the value in the min or max gauge, then enable the custom range checkbutton, labels, gauges. lastly call AdjustSetting -range to redraw / recalculate isosurfaces and update the legend labels with the new min and max values.